+++ src/neo4j_graphrag/experimental/components/types.py
@@ -105,21 +105,32 @@ class ResolutionStats(DataModel):
     number_of_created_nodes: Optional[int] = None
 
 
-CHUNK_NODE_LABEL = "Chunk"
-DOCUMENT_NODE_LABEL = "Document"
-NEXT_CHUNK_RELATIONSHIP_TYPE = "NEXT_CHUNK"
-CHUNK_TO_DOCUMENT_RELATIONSHIP_TYPE = "FROM_DOCUMENT"
-NODE_TO_CHUNK_RELATIONSHIP_TYPE = "FROM_CHUNK"
+DEFAULT_DOCUMENT_NODE_LABEL = "Document"
+DEFAULT_CHUNK_NODE_LABEL = "Chunk"
+DEFAULT_CHUNK_TO_DOCUMENT_RELATIONSHIP_TYPE = "FROM_DOCUMENT"
+DEFAULT_NEXT_CHUNK_RELATIONSHIP_TYPE = "NEXT_CHUNK"
+DEFAULT_NODE_TO_CHUNK_RELATIONSHIP_TYPE = "FROM_CHUNK"
+DEFAULT_CHUNK_EMBEDDING_PROPERTY = "embedding"
 
 
 class LexicalGraphConfig(BaseModel):
     id_prefix: str = ""
-    document_node_label: str = DOCUMENT_NODE_LABEL
-    chunk_node_label: str = CHUNK_NODE_LABEL
-    chunk_to_document_relationship_type: str = CHUNK_TO_DOCUMENT_RELATIONSHIP_TYPE
-    next_chunk_relationship_type: str = NEXT_CHUNK_RELATIONSHIP_TYPE
-    node_to_chunk_relationship_type: str = NODE_TO_CHUNK_RELATIONSHIP_TYPE
-
-    chunk_embedding_property: str = "embedding"
+    document_node_label: str = DEFAULT_DOCUMENT_NODE_LABEL
+    chunk_node_label: str = DEFAULT_CHUNK_NODE_LABEL
+    chunk_to_document_relationship_type: str = (
+        DEFAULT_CHUNK_TO_DOCUMENT_RELATIONSHIP_TYPE
+    )
+    next_chunk_relationship_type: str = DEFAULT_NEXT_CHUNK_RELATIONSHIP_TYPE
+    node_to_chunk_relationship_type: str = DEFAULT_NODE_TO_CHUNK_RELATIONSHIP_TYPE
+    chunk_embedding_property: str = DEFAULT_CHUNK_EMBEDDING_PROPERTY
     chunk_text_property: str = "text"
     chunk_index_property: str = "index"
+
+    @property
+    def lexical_graph_node_labels(self) -> tuple[str, str]:
+        return (self.document_node_label, self.chunk_node_label)
+
+
+class GraphResult(DataModel):
+    graph: Neo4jGraph
+    config: LexicalGraphConfig